refactor: refactored header-layout and image layouts files
